+++ tests/panels/test_template.py
@@ -1,6 +1,5 @@
 from unittest import expectedFailure
 
-import django
 from django.contrib.auth.models import User
 from django.template import Context, RequestContext, Template
 from django.test import override_settings
@@ -56,10 +55,7 @@ class TemplatesPanelTestCase(BaseTestCase):
 
         User.objects.create(username="admin")
         bad_repr = TemplateReprForm()
-        if django.VERSION < (5,):
-            t = Template("<table>{{ bad_repr }}</table>")
-        else:
-            t = Template("{{ bad_repr }}")
+        t = Template("{{ bad_repr }}")
         c = Context({"bad_repr": bad_repr})
         html = t.render(c)
         self.assertIsNotNone(html)